% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_info.R
\name{cluster_info}
\alias{cluster_info}
\title{cluster_info}
\usage{
cluster_info(df)
}
\value{
A list (invisibly) containing:
\describe{
\item{cluster_amount}{Number of unique clusters at the last time step}
\item{cluster_sizes}{Table of cluster sizes at the last time step}
\item{cluster_stats}{Data frame with statistics for each cluster }
\item{cluster_stats_full}{Data frame that contains the cluster statistics for every time step}
\item{cluster_history}{Data frame showing each member's cluster assignment for each member and time step}
\item{cluster_history_full}{Data frame showing original cluster assignments for each member and time step}
\item{original_data}{Dataframe of the original data with Timeindex, date_and_time, Member, and Measure}
}
}
\description{
This function gives a summary of the clustered dataframe of define_cluster

#'@param df A dataframe which was created using the define_cluster function
}
\details{
When calling the function, the following elements are printed to the console:
\itemize{
\item cluster_amount
\item cluster_sizes
\item cluster_stats
}
}
\examples{
cluster_info(data)
cluster_info(data)$cluster_history
}
