% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cluster.R
\name{define_cluster}
\alias{define_cluster}
\title{define_cluster}
\usage{
define_cluster(
  data,
  k = 2,
  limit = function(km, subdata, a = 4) max(diff(km$centers)) > a,
  max_cluster = 8
)
}
\arguments{
\item{data}{A dataframe created by the function get_data

Or a dataframe that at least includes : a clomun named Timeindex and at least 1 column with numeric data}

\item{k}{Initial number of clusters that should be created when the limit is reached.}

\item{limit}{A numeric value or a function that defines the limit that has to be reached for the clustering

If a number is provided, it is treated as the limit for \code{max(diff(km$centers))}.

Default is \code{max(diff(km$centers)) > 4}

If given a custom limit function the function should return TRUE if the clustering should be done}

\item{max_cluster}{Maximum allowed number of clusters.}
}
\value{
A data frame with additional columns:
\describe{
\item{Cluster}{Raw Clustersequence (eg. 1.1.1.1)}
\item{Clusterlabel}{Clusterlabel that only counts splits as a new cluster}
\item{betweenss, totss, totwithinss, betweensstotss, rsquarred}{Cluster statistics per time step}
\item{Measure}{Original numeric measurement}
\item{Member}{Different ensemble member}
\item{Timeindex}{Time step index}
\item{date_and_time}{Date and time matching the time steps}
}
}
\description{
Define clusters in ensemble forecast data

This function performs dynamic clustering on ensemble forecast data.

It applies k-means clustering for each time step where a certain limit is reached.
And allocates the Clusters to their Members. The next Clustering only happens with members inside the current cluster.
Clusterlabels are given by the size of the clusters - the Cluster with the most members gets 1,second most 2,...

The Labels of the cluster give a hint about their development over time.
}
\examples{
clustered_df <- define_cluster(data,k=2,limit=4)

}
