% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data()}
\usage{
get_data(
  date = gsub("-", "", Sys.Date()),
  time = "00",
  lat = 282,
  lon = 28,
  variable = "tmpprs",
  pressr = 2
)
}
\arguments{
\item{date}{string that defines the date of the forecast

has to be in the format "yyyymmdd"}

\item{time}{string that defines the time of the forecast

only "00","06","12","18" are valid inputs

defines from which time the forecast should be}

\item{lat}{Integer latitude of the desired place with .5 accuracy

ranges from 0-360}

\item{lon}{Integer longitude of the desired place with .5 accuracy

ranges from 0-719}

\item{variable}{String that defines wich environmental variable should be fetched

some examples: tmpprs,hgtprs,ugrdprs}

\item{pressr}{Integer pressurelevels for the data

ranges from 0-11 - higher values = higher in the atmosphere}
}
\value{
A Dataframe with environmental data for 31 Members of a Ensembleforcast from NOMADS depending on the given parameters
}
\description{
This Function fetches data from https://nomads.ncep.noaa.gov, transforms it and loads it into a dataframe.
}
\examples{
x <- get_data("20250825",lat = 280,lon = 30)
y <- get_data(variable = "hgtprs",pressr = 0)
}
