% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster.R
\name{plot_cluster}
\alias{plot_cluster}
\title{create plots for clustered data}
\usage{
plot_cluster(
  data,
  col_palette = c("#67FF00", "#97FF00", "#FFF000", "#FFB556", "#FF7133", "#FF0000",
    "#5700A4", "#2100FF"),
  return_data = FALSE,
  linesize = 1.2,
  linealpha = 1,
  points = list(),
  fade = TRUE,
  fade_till_end = FALSE,
  legend = TRUE,
  steps = 1,
  background = "grey37",
  axis_bg = "grey37",
  grid_col = "white",
  xaxis_text = list(color = "black", size = 15),
  yaxis_text = list(color = "black", size = 15),
  break_times = "2 day"
)
}
\arguments{
\item{data}{A dataframe created by the define_cluster function}

\item{col_palette}{Character vector of colors to use for clusters.}

\item{return_data}{Logical. If TRUE, returns a list containing the plot and the processed data.}

\item{linesize}{Numeric. defines which size the line should have}

\item{linealpha}{Numeric between 0 and 1. Transparency of the cluster lines.}

\item{points}{List. Additional arguments for adding points (e.g. size).}

\item{fade}{Logical. If TRUE a color of a cluster fades to their next}

\item{fade_till_end}{Logical. If TRUE, cluster colors fade until the end of the time series.}

\item{legend}{Logical. If TRUE, adds a legend containing which clusters have which color}

\item{steps}{Numeric. Number of extra steps between 2 time steps, creates smoother fade coloring}

\item{background}{Character. Background color of the plot.}

\item{axis_bg}{Character. Background color of the axis.}

\item{grid_col}{Character. Color of grid lines in the plot.}

\item{xaxis_text}{List of parameters passed to \code{element_text()} for x-axis text (e.g., color, size).}

\item{yaxis_text}{List of parameters passed to \code{element_text()} for y-axis text.}

\item{break_times}{Character. Breaks for the x-axis in \code{scale_x_datetime}.}
}
\value{
If \code{return_data = FALSE} (default), returns a \code{ggplot} object.
If \code{return_data = TRUE}, returns a list containing the plot and the data needed for the plot
}
\description{
Plot cluster evolution for ensemble forecast data.

This function visualizes the evolution of clusters over time for
ensemble forecast data.

It allows customization of colors, line widths,
points, fading effects, and axis appearance. Optionally, the processed
data can also be returned.
}
\examples{
plot_cluster(
  data,
  col_palette = c("#FFFB00","#FF7E20","#FF0000","#ED0DFF"),
  linesize = 1,
  fade_till_end = T,
  break_times = "2 day",
  points= list(size = 1.5)

)
}
