% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdata.R
\docType{data}
\name{tempdata}
\alias{tempdata}
\title{Example dataset: tempdata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65 rows and 33 columns.
}
\source{
https://nomads.ncep.noaa.gov/dods/gefs/gefs20250828/gefs_pgrb2ap5_all_00z.ascii?tmpprs[0:30][0:64][2][282][28]
}
\usage{
tempdata
}
\description{
Data with temperature data forecast from 28.08.25 (tmpprs)
}
\details{
Is a dataframe generated by get_data
}
\examples{
data("tempdata", package = "Enclust")
head(tempdata)
}
\keyword{datasets}
