% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winddata.R
\docType{data}
\name{winddata}
\alias{winddata}
\title{Example dataset: winddata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65 rows and 33 columns.
}
\source{
https://nomads.ncep.noaa.gov/dods/gefs/gefs20250828/gefs_pgrb2ap5_all_00z.ascii?hgtprs[0:30][0:64][2][282][28]
}
\usage{
winddata
}
\description{
Data with windspeed data forecast from 28.08.25 (hgtprs)
}
\details{
Is a dataframe generated by get_data
}
\examples{
data("winddata", package = "Enclust")
head(winddata)
}
\keyword{datasets}
